# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Dalikamata API
  description: |
    HTTP query API for raw entity access over the Dalikamata CD metrics platform.

    Each endpoint accepts both **GET** (URL query parameters) and **POST**
    (a full `Query` JSON body).  The GET interface is a convenient shorthand;
    the POST interface gives access to the complete query DSL including boolean
    filter trees and server-side aggregations.

    ### Aggregation modes
    Set `"aggs_only": true` in a POST body to suppress entity hits and receive
    only aggregation results (`AggregationResponse`).

    Omit `aggs_only` (or set it to `false`) and include a non-empty `"aggs"` map
    to receive both paginated hits **and** aggregations in a single response
    (`HitsResponse` with the optional `aggregations` field populated).

    ### GET filter parameters
    Filters are expressed as query parameters with a `filter.` prefix:

    | Pattern | Filter type | Example |
    |---|---|---|
    | `filter.<field>=<value>` | term (exact match) | `filter.state=MERGED` |
    | `filter.<field>=<v1>&filter.<field>=<v2>` | terms (any of) | `filter.status=SUCCESS&filter.status=FAILURE` |
    | `filter.<field>.gte=<value>` | range lower bound (inclusive) | `filter.started_at.gte=2024-01-01T00:00:00Z` |
    | `filter.<field>.lte=<value>` | range upper bound (inclusive) | `filter.started_at.lte=2024-12-31T23:59:59Z` |
    | `filter.<field>.gt=<value>` | range lower bound (exclusive) | |
...

...

...
  version: "1"
...
servers:
...
security: []
...
paths:
...
  /api/v1/ownership:
    get:
      operationId: ownershipDiagnostics
      summary: Ownership resolution report
      description: >
        Returns one entry per known Workflow showing how it resolves (or fails
        to resolve) via the chain Workflow.RepoID → Component.RepoIDs →
        Component.TeamName. Use this to diagnose why a workflow shows
        team_name="unknown" in metrics or query results.
      tags: [platform]
      responses:
        "200":
          description: Ownership resolution report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OwnershipDiagnostics"
        "504":
          $ref: "#/components/responses/GatewayTimeout"

# ── Reusable components ───────────────────────────────────────────────────────
components:
...
  responses:
...
    GatewayTimeout:
      description: Query exceeded the configured timeout
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Entity names ─────────────────────────────────────────────────────────
...
    OwnershipDiagnostics:
      type: object
      required: [workflow_id, repo_id, component_name, team_name, reason]
      properties:
        workflow_id:
          type: string
        repo_id:
          type: string
          description: Empty if the workflow has no RepoID set (extraction failed at ingest).
        component_name:
          type: string
          description: Empty if no component claims the repo.
        team_name:
          type: string
          description: Empty if the component has no team mapping.
        reason:
          type: string
          enum: [ok, missing_repo_id, no_component_for_repo, no_team_for_component]
          description: >
            ok — chain resolved fully.
            missing_repo_id — Workflow.RepoID is empty (Jenkins BuildData not populated).
            no_component_for_repo — no Component lists this RepoID.
            no_team_for_component — Component found but TeamName is missing.